fix: type of parameter

+++ src/main/java/com/dexdev/loansimulation/modules/simulation/controllers/dtos/CreateSimulationRequest.java
@@ -11,7 +11,7 @@ public class CreateSimulationRequest implements Serializable {
     public Integer clientId;
     @NotNull
     @Min(1000)
-    public Integer amount;
+    public double amount;
     @NotNull
     @Min(1)
     @Max(12)
